pyproject.toml
[project]
name = "face-to-age"
version = "0.1.0"
description = ""
authors = [
    {name = "Max_Arkhipov",email = "maarchipov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "pytorch-lightning (>=2.6.0,<3.0.0)",
    "torchmetrics (>=1.8.2,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "mlflow (>=3.8.0,<4.0.0)",
    "dvc[gdrive] (>=3.65.0,<4.0.0)",
    "torch (>=2.9.1,<3.0.0)",
    "torchvision (>=0.24.1,<0.25.0)",
    "pyopenssl (>=24.0.0)",
    "cryptography (>=42.0.0)",
    "lightning (>=2.6.0,<3.0.0)",
    "ipywidgets (>=8.1.8,<9.0.0)",
    "fire (>=0.7.1,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "ruff (>=0.14.10,<0.15.0)"
]

[tool.ruff]
line-length = 100
format = { docstring-code-format = true }

[tool.ruff.lint]
select = [
    "I", "E", "W", "F"
]

pycodestyle = { max-doc-length = 105 }
